Implement App Authentication with JWT

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
   try {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json({ goals: goals });
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -24,6 +25,7 @@ const setGoal = asyncHandler(async (req, res) => {
   try {
     const goal = await Goal.create({
       text: req.body.text,
+      user: req.user.id,
     });
     res.status(200).json({ goal: goal });
   } catch (error) {
@@ -40,6 +42,17 @@ const updateGoal = asyncHandler(async (req, res) => {
     if (!goal) {
       res.status(404);
       throw new Error("Goal not found");
+    }
+    // Check for user
+    const user = await User.findById(req.user.id);
+    if (!user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error("User not authorized");
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
@@ -60,6 +73,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
       res.status(404);
       throw new Error("Goal not found");
     }
+    // Check for user
+    const user = await User.findById(req.user.id);
+    if (!user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
+     
     await goal.deleteOne();
     res.status(200).json({ id: req.params.id });
   } catch (error) {
